Use CombinePDF and test behavior for WriteToDocument spec
